refactor(desktop): unify comment + document push-on-publish via shared hook
Extract the push-to-servers plumbing shared between comment publishing
and the unified document editor into a usePushAfterPublish hook. Both
call sites collapse to a single call, and the pushDocument actor no
longer needs separate refs for pushResource and pushOnPublish.

The compact PushedToast now surfaces progress as "(done/total)" during
loading once destination hosts are known. The comment flow switches
from the verbose per-host PublishedToast to this compact variant for
consistency with the document editor.
