+++ frontend/apps/desktop/src/components/commenting.tsx
@@ -1,6 +1,5 @@
 import {useCommentDraft} from '@/models/comments'
-import {usePushResource} from '@/models/documents'
-import {usePushOnPublish} from '@/models/gateway-settings'
+import {usePushAfterPublish} from '@/models/push-after-publish'
 import {useSelectedAccount, useSelectedAccountId} from '@/selected-account'
 import {client} from '@/trpc'
 import {handleDragMedia} from '@/utils/media-drag'
@@ -25,10 +24,7 @@ import {invalidateQueries} from '@shm/shared/models/query-client'
 import {applyOptimisticComment, buildOptimisticComment, navigateToComment} from '@shm/shared/optimistic-comment'
 import {useUniversalClient} from '@shm/shared/routing'
 import {useNavRoute} from '@shm/shared/utils/navigation'
-import {writeableStateStream} from '@shm/shared/utils/stream'
 import {Button} from '@shm/ui/button'
-import {PublishedToast, PushResourceStatus} from '@shm/ui/push-toast'
-import {toast} from '@shm/ui/toast'
 import {Tooltip} from '@shm/ui/tooltip'
 import {useMutation} from '@tanstack/react-query'
 import {Check, SendHorizonal, X} from 'lucide-react'
@@ -70,8 +66,7 @@ function _CommentBox(props: {
   const account = useSelectedAccount()
   const selectedAccountId = useSelectedAccountId()
   const targetEntity = useResource(docId)
-  const pushResource = usePushResource()
-  const pushOnPublish = usePushOnPublish()
+  const pushAfterPublish = usePushAfterPublish()
   const {getSigner, publish} = useUniversalClient()
   const route = useNavRoute()
   const navigate = useNavigate('replace')
@@ -217,15 +212,7 @@ function _CommentBox(props: {
       invalidateQueries([queryKeys.BLOCK_DISCUSSIONS])
       invalidateQueries([queryKeys.ACTIVITY_FEED])
 
-      if (pushOnPublish.data !== 'never') {
-        const [setPushStatus, pushStatus] = writeableStateStream<PushResourceStatus | null>(null)
-        const pushPromise = pushResource(docId, undefined, setPushStatus)
-        toast.promise(pushPromise, {
-          loading: <PublishedToast pushStatus={pushStatus} status="loading" />,
-          success: <PublishedToast pushStatus={pushStatus} status="success" />,
-          error: (error) => <PublishedToast pushStatus={pushStatus} status="error" errorMessage={error.message} />,
-        })
-      }
+      pushAfterPublish(docId)
     },
   })
